+++ lib/fs/irmin_fs.ml
@@ -32,7 +32,7 @@ module type IO = sig
   val remove: string -> unit Lwt.t
   val rec_files: string -> string list Lwt.t
   val read_file: string -> Cstruct.t Lwt.t
-  val write_file: string -> Cstruct.t -> unit Lwt.t
+  val write_file: string -> ?temp_dir:string -> Cstruct.t -> unit Lwt.t
 end
 
 (* ~path *)
@@ -166,6 +166,10 @@ module RW_ext (IO: IO) (S: Config) (K: Irmin.Hum.S) (V: Tc.S0) = struct
       read t key >>= fun value ->
       return (W.watch t.w key value)
     )
+
+  let watch_all t =
+    W.listen_dir t.w t.path ~key:key_of_file ~value:(read t);
+    W.watch_all t.w
 
 end