+++ lib/ir_watch.ml
@@ -27,6 +27,7 @@ module type S = sig
   val create: unit -> t
   val clear: t -> unit
   val watch: t -> key -> value option -> value option Lwt_stream.t
+  val watch_all: t -> (key * value option) Lwt_stream.t
   val listen_dir: t -> string
     -> key:(string -> key option)
     -> value:(key -> value option Lwt.t)
@@ -34,10 +35,17 @@ module type S = sig
 end
 
 let listen_dir_hook =
-  ref (fun _dir _fn -> ())
+  ref (fun _dir _fn ->
+      Log.error "Listen hook not set!";
+      assert false
+    )
 
 let set_listen_dir_hook fn =
   listen_dir_hook := fn
+
+let id =
+  let c = ref 0 in
+  fun () -> incr c; !c
 
 module Make (K: Tc.S0) (V: Tc.S0) = struct
 
@@ -45,58 +53,111 @@ module Make (K: Tc.S0) (V: Tc.S0) = struct
   type value = V.t
   module OV = Tc.Option(V)
 
-  type t = (K.t, (int * value option * (value option option -> unit)) list) Hashtbl.t
+  type key_notifier = int * value option * (value option option -> unit)
+  type all_notifier = int * ((key * value option) option -> unit)
+
+  type t = {
+    id: int;
+    keys: (K.t,  key_notifier list) Hashtbl.t;
+    mutable all: all_notifier list;
+  }
+
+  let to_string t =
+    Printf.sprintf "[%d|%s|%d]"
+      t.id
+      (String.concat " "
+         (let l = Hashtbl.fold (fun k _ l -> k :: l) t.keys [] in
+          List.map (Tc.show (module K)) l))
+      (List.length t.all)
 
   let create () =
-    Hashtbl.create 42
+    { id = id (); keys = Hashtbl.create 42; all = [] }
 
   let clear t =
-    Hashtbl.clear t
+    Hashtbl.clear t.keys;
+    t.all <- []
 
-  let unwatch t key id =
+  let close_key_notofier = fun (_, _, f) -> f None
+  let close_all_notifier = fun (_, f) -> f None
+
+  let unwatch_keys (t:t) key id =
     let ws =
-      try Hashtbl.find t key
+      try Hashtbl.find t.keys key
       with Not_found -> []
     in
     let ws = List.filter (fun (x,_,_) -> x <> id) ws in
     (* close the clients *)
-    List.iter (fun (_, _, f) -> f None) ws;
+    List.iter close_key_notofier ws;
     match ws with
-    | [] -> Hashtbl.remove t key
-    | ws -> Hashtbl.replace t key ws
+    | [] -> Hashtbl.remove t.keys key
+    | ws -> Hashtbl.replace t.keys key ws
 
-  let notify t key value =
-    Log.debug "notify %a" force (show (module K) key);
+  let notify_keys (t:t) key value =
     try
-      let ws = Hashtbl.find t key in
-      let ws = List.map (fun (id, old_value, f as w) ->
-          if not (OV.equal old_value value) then (
-            Log.debug "firing watch %a:%d" force (show (module K) key) id;
-            try f (Some value); (id, value, f)
-            with e ->
-              unwatch t key id;
-              raise e
-          ) else w
-        ) ws
+      let ws = Hashtbl.find t.keys key in
+      let ws =
+        List.fold_left (fun acc (id, old_value, f as w) ->
+            if not (OV.equal old_value value) then (
+              Log.debug "firing key-watch id=%d.%d %a"
+                t.id id force (show (module K) key);
+              try f (Some value); (id, value, f) :: acc
+              with e ->
+                Log.error "notify-key: id=%d.%d %s" t.id id
+                  (Printexc.to_string e);
+                unwatch_keys t key id;
+                acc
+            ) else w :: acc
+          ) [] ws
+        |> List.rev
       in
-      Hashtbl.replace t key ws
+      Hashtbl.replace t.keys key ws
     with Not_found ->
       ()
 
-  let id =
-    let c = ref 0 in
-    fun () -> incr c; !c
+  let unwatch_all (t:t) id =
+    let to_close, all = List.partition (fun (x, _) -> x = id) t.all in
+    List.iter close_all_notifier to_close;
+    t.all <- all
+
+  let notify_all (t:t) key value =
+    let all =
+      List.fold_left (fun acc (id, f) ->
+          Log.debug "firing all-watch id=%d.%d %a" t.id id force
+            (show (module K) key);
+          try f (Some (key, value)); (id, f) :: acc
+          with e ->
+            Log.debug "notify-all: id=%d.%d %s" t.id id (Printexc.to_string e);
+            unwatch_all t id;
+            acc
+        ) [] t.all
+      |> List.rev
+    in
+    t.all <- all
+
+  let notify t key value =
+    Log.debug "notify %s %a" (to_string t) force
+      (show (module K) key);
+    notify_keys t key value;
+    notify_all t key value
 
   let watch (t:t) key value =
-    Log.debug "watch %a" force (show (module K) key);
+    Log.debug "watch %s %a" (to_string t) force (show (module K) key);
     let stream, push = Lwt_stream.create () in
     let id = id () in
-    Ir_misc.hashtbl_add_multi t key (id, value, push);
+    Ir_misc.hashtbl_add_multi t.keys key (id, value, push);
     stream
 
-  let listen_dir t dir ~key ~value =
-    !listen_dir_hook dir (fun file ->
-        Log.debug "listen_dir_hook: %s" file;
+  let watch_all (t:t) =
+    Log.debug "watch all %s" (to_string t);
+    let stream, push = Lwt_stream.create () in
+    let id = id () in
+    t.all <- (id, push) :: t.all;
+    stream
+
+  let listen_dir (t:t) dir ~key ~value =
+    Log.debug "Add a listen hook for %s" (to_string t);
+    !listen_dir_hook t.id dir (fun file ->
+        Log.debug "listen_dir_hook: %s %s" (to_string t) file;
         match key file with
         | None     -> return_unit
         | Some key ->