+++ lib/irmin.ml
@@ -105,6 +105,7 @@ type ('a, 'k, 'v) t = {
   update: 'k -> 'v -> unit Lwt.t;
   remove: 'k -> unit Lwt.t;
   watch: 'k -> 'v option Lwt_stream.t;
+  watch_all: unit -> ('k * 'v option) Lwt_stream.t;
   list: 'k -> 'k list Lwt.t;
   remove_rec: 'k -> unit Lwt.t;
   extend: ('a, 'k, 'v) ext;
@@ -132,17 +133,18 @@ and ('k, 'v) bc = {
   update_head: Hash.SHA1.t -> unit Lwt.t;
   merge_head: ?max_depth:int -> ?n:int -> Hash.SHA1.t -> unit Merge.result Lwt.t;
   merge_head_exn: ?max_depth:int -> ?n:int -> Hash.SHA1.t -> unit Lwt.t;
-  watch_head: 'k -> ('k * Hash.SHA1.t) Lwt_stream.t;
+  watch_head: 'k -> ('k * Hash.SHA1.t option) Lwt_stream.t;
+  watch_tags: unit -> (string * Hash.SHA1.t option) Lwt_stream.t;
   clone: 'm. ('m -> task) -> string -> [`Ok of ('m -> ([`BC], 'k, 'v) t) | `Duplicated_tag] Lwt.t;
   clone_force: 'm. ('m -> task) -> string -> ('m -> ([`BC], 'k, 'v) t) Lwt.t;
   merge: ?max_depth:int -> ?n:int -> into:([`BC], 'k, 'v) t ->
     unit Merge.result Lwt.t;
   merge_exn: ?max_depth:int -> ?n:int -> into:([`BC], 'k, 'v) t -> unit Lwt.t;
-  lca: ?max_depth:int -> ?n:int -> ([`BC], 'k, 'v) t ->
+  lcas: ?max_depth:int -> ?n:int -> ([`BC], 'k, 'v) t ->
     [`Ok of Hash.SHA1.t list | `Too_many_lcas | `Max_depth_reached] Lwt.t;
-  lca_tag: ?max_depth:int -> ?n:int -> string ->
+  lcas_tag: ?max_depth:int -> ?n:int -> string ->
     [`Ok of Hash.SHA1.t list | `Too_many_lcas | `Max_depth_reached] Lwt.t;
-  lca_head: ?max_depth:int -> ?n:int -> Hash.SHA1.t ->
+  lcas_head: ?max_depth:int -> ?n:int -> Hash.SHA1.t ->
   [`Ok of Hash.SHA1.t list | `Too_many_lcas | `Max_depth_reached] Lwt.t;
   task_of_head: Hash.SHA1.t -> task Lwt.t;
   remote_basic: unit -> remote;
@@ -162,6 +164,7 @@ let iter t = t.iter
 let update t = t.update
 let remove t = t.remove
 let watch t = t.watch
+let watch_all t = t.watch_all ()
 let list t = t.list
 let remove_rec t = t.remove_rec
 
@@ -185,16 +188,17 @@ let update_head t = bc t (function BC t -> t.update_head)
 let merge_head t = bc t (function BC t -> t.merge_head)
 let merge_head_exn t = bc t (function BC t -> t.merge_head_exn)
 let watch_head t = bc t (function BC t -> t.watch_head)
+let watch_tags t = bc t (function BC t -> t.watch_tags ())
 let clone task t = bc t (function BC t -> t.clone task)
 let clone_force task t = bc t (function BC t -> t.clone_force task)
 let merge a ?max_depth ?n t ~into =
   bc (t a) (function BC t -> t.merge ?max_depth ?n ~into:(into a))
 let merge_exn a ?max_depth ?n t ~into =
   bc (t a) (function BC t -> t.merge_exn ?max_depth ?n ~into:(into a))
-let lca a ?max_depth ?n t1 t2 =
-  bc (t1 a) (function BC t1 -> t1.lca ?max_depth ?n (t2 a))
-let lca_tag t = bc t (function BC t -> t.lca_tag)
-let lca_head t = bc t (function BC t -> t.lca_head)
+let lcas a ?max_depth ?n t1 t2 =
+  bc (t1 a) (function BC t1 -> t1.lcas ?max_depth ?n (t2 a))
+let lcas_tag t = bc t (function BC t -> t.lcas_tag)
+let lcas_head t = bc t (function BC t -> t.lcas_head)
 let task_of_head t = bc t (function BC t -> t.task_of_head)
 
 (* sync *)
@@ -219,6 +223,7 @@ let pack_hrw (type x) (type k) (type v)
   update = M.update t;
   remove = M.remove t;
   watch = M.watch t;
+  watch_all = (fun () -> M.watch_all t);
   list = M.list t;
   remove_rec = M.remove_rec t;
   extend = HRW;
@@ -254,6 +259,7 @@ let pack_s (type x) (type k) (type v)
         merge_head = M.merge_head t;
         merge_head_exn = M.merge_head_exn t;
         watch_head = M.watch_head t;
+        watch_tags = (fun () -> M.watch_tags t);
         clone = (fun task tag ->
             M.clone task t tag >>= function
             | `Ok x           -> Lwt.return (`Ok (fun a -> aux (x a)))
@@ -267,14 +273,14 @@ let pack_s (type x) (type k) (type v)
         merge_exn = (fun ?max_depth ?n ~into ->
             merge t ?max_depth ?n ~into >>= Ir_merge.exn
           );
-        lca = (fun ?max_depth ?n i ->
+        lcas = (fun ?max_depth ?n i ->
             let BC i = i.extend in
             match i.branch () with
-            | `Tag tag -> M.lca_tag t ?max_depth ?n tag
-            | `Head h  -> M.lca_head t ?max_depth ?n h
+            | `Tag tag -> M.lcas_tag t ?max_depth ?n tag
+            | `Head h  -> M.lcas_head t ?max_depth ?n h
           );
-        lca_tag = M.lca_tag t;
-        lca_head = M.lca_head t;
+        lcas_tag = M.lcas_tag t;
+        lcas_head = M.lcas_head t;
         task_of_head = M.task_of_head t;
         remote_basic = (fun () -> remote_store (module M) t);
         fetch = S.fetch t;
@@ -305,11 +311,9 @@ let of_head (type a) (type b) (t: (a, b) basic) config task h =
   T.of_head config task h >>= fun t ->
   return (pack_s (module T) t)
 
-type 'a proj = { f: 't . (module S with type t = 't) -> 't -> 'a }
-
-let with_store (type a) (type b): ([`BC], a, b) t -> 'a proj -> 'a =
-  fun t f -> match t.extend with
-    | BC { pack = E ((module M), t); _ } -> f.f (module M) t
+let impl (type a) (type b): ([`BC], a, b) t -> (a, b) basic =
+  fun t -> match t.extend with
+    | BC { pack = E ((module M), _); _ } -> (module M)
 
 let with_hrw_view (type a) (type b)
     (t :([`BC], a, b) t) ?path strat (ops: ([`HRW], a, b) t -> unit Lwt.t) =