+++ CHANGES.md
@@ -1,4 +1,7 @@
 ## 0.9.3
+* Add post-commit hooks for the HTTP server.
+* Add `Irmin.watch_tags` to monitor tag creation and desctructions.
+* Fix `Irmin.push`
 * Add `Irmin.with_hrw_view` to easily use transactions.
 * Add a phantom type to `Irmin.t` to denote the store capabilities
   read-only, read-write or branch-consistent.